refactor: refactor import statement - using absolute path
