+++ .eslintrc.json
@@ -66,7 +66,6 @@
       }
     ],
     "react-hooks/rules-of-hooks": "error"
-    // "sort-keys-fix/sort-keys-fix": "warn"
   },
   "settings": {
     "import/parsers": {
