+++ package.json
@@ -38,6 +38,7 @@
     "react-router": "^6.3.0",
     "react-router-dom": "^6.3.0",
     "styled-components": "^5.3.5",
+    "vite-tsconfig-paths": "^3.5.1",
     "yup": "^0.32.11"
   },
   "devDependencies": {
