+++ src/__tests__/hooks/useDidUpdateEffect.test.ts
@@ -1,6 +1,6 @@
 import { renderHook } from '@testing-library/react'
 
-import useDidUpdateEffect from '../../hooks/useDidUpdateEffect'
+import useDidUpdateEffect from './../../hooks/useDidUpdateEffect'
 
 test('useUpdateEffect simulates componentDidUpdate', () => {
   const effect = jest.fn()
