+++ src/components/App.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 
-import { useAuth } from '../hooks/useAuth'
+import { useAuth } from 'hooks/useAuth'
 
 import { FullPageSpinner } from './Layout/FullPageSpinner'
