+++ src/components/Authenticated/index.tsx
@@ -2,10 +2,10 @@ import React from 'react'
 import { ErrorBoundary } from 'react-error-boundary'
 import { Routes, Route, Navigate } from 'react-router-dom'
 
-import { useAuth } from '../../hooks/useAuth'
-import { ErrorFallback } from '../Layout/ErrorFallback'
-import { FullPageErrorFallback } from '../Layout/FullPageErrorFallback'
-import NavHeader from '../Layout/NavHeader'
+import { ErrorFallback } from 'components/Layout/ErrorFallback'
+import { FullPageErrorFallback } from 'components/Layout/FullPageErrorFallback'
+import NavHeader from 'components/Layout/NavHeader'
+import { useAuth } from 'hooks/useAuth'
 
 import PetDetail from './Pet/PetDetail'
 import PetList from './Pet/PetList'
