--- src/components/Common/FileUpload.tsx
@@ -1,7 +1,0 @@
-import React from 'react'
-
-const FileUpload = () => {
-  return <div>FileUpload</div>
-}
-
-export default FileUpload
