+++ src/components/Layout/FullPageErrorFallback.tsx
@@ -1,4 +1,4 @@
-import * as colors from '../../constants/colors'
+import * as colors from 'constants/colors'
 
 function FullPageErrorFallback(error: { error: Error }) {
   return (
