+++ src/database/index.ts
@@ -1,7 +1,7 @@
 import type { Table } from 'dexie'
 import Dexie from 'dexie'
 
-import type { PetDexieModel, OwnerDexieModel } from '../src/types/types'
+import type { PetDexieModel, OwnerDexieModel } from '../types/types'
 
 export class MySubClassedDexie extends Dexie {
   pet!: Table<PetDexieModel>
