+++ src/jest/setupTests.js
@@ -2,7 +2,7 @@ import '@testing-library/jest-dom'
 
 // Polyfill "window.fetch" used in the React component.
 import 'whatwg-fetch'
-import { server } from './mocks/server'
+import { server } from '../mocks/server'
 
 beforeAll(() => {
   server.listen()
